Removed CellRoot
Added comments to patches
