+++ NitroxModel/NitroxModel.csproj
@@ -223,7 +223,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DataStructures\CellRoot.cs" />
     <Compile Include="Core\IAutoFacRegistrar.cs" />
     <Compile Include="Core\NitroxServiceLocator.cs" />
     <Compile Include="DataStructures\GameLogic\Entities\UwePrefab.cs" />
