+++ NitroxPatcher/Patches/Dynamic/CrashHome_Spawn_Patch.cs
@@ -9,7 +9,7 @@ namespace NitroxPatcher.Patches.Dynamic
         public static readonly Type TARGET_CLASS = typeof(CrashHome);
         public static readonly MethodInfo TARGET_METHOD = TARGET_CLASS.GetMethod("Spawn", BindingFlags.NonPublic | BindingFlags.Instance);
 
-        public static bool Prefix()
+        public static bool Prefix() // Disables Crashfish automatic spawning on the client
         {
             return false;
         }
