+++ NitroxPatcher/Patches/Persistent/ProtobufSerializer_Serialize_Patch.cs
@@ -3,24 +3,24 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text;
-using System.IO;
 using Harmony;
 using ProtoBuf;
+using System.IO;
 using NitroxClient.Helpers;
 
 namespace NitroxPatcher.Patches.Persistent
 {
-    public class ProtobufSerializerPrecompiled_Deserialize_Patch : NitroxPatch, IPersistentPatch
+    public class ProtobufSerializer_Serialize_Patch : NitroxPatch, IPersistentPatch
     {
         static Type TARGET_TYPE = typeof(ProtobufSerializer);
-        static MethodInfo TARGET_METHOD = TARGET_TYPE.GetMethod("Deserialize", BindingFlags.Instance | BindingFlags.NonPublic);
+        static MethodInfo TARGET_METHOD = TARGET_TYPE.GetMethod("Serialize", BindingFlags.Instance | BindingFlags.NonPublic);
 
-        public static bool Prefix(Stream stream, object target, Type type)
+        public static bool Prefix(Stream stream, object source, Type type)
         {
             int key;
             if (NitroxProtobufSerializer.Main.NitroxTypes.TryGetValue(type, out key))
             {
-                NitroxProtobufSerializer.Main.Deserialize(stream, target, type);
+                NitroxProtobufSerializer.Main.Serialize(stream, source);
                 return false;
             }
